// src/models/contactTranslations.ts
export interface ContactTranslations {
  contactDescription: string;
  contactInvitation: string;
  errorName: string;
  name: string;
  email: string;
  errorEmail: string;
  phone: string;
  errorPhone: string;
  message: string;
  errorMessage: string;
  contactUsButton: string;
  thanksForContacting: string;
  contactSoon: string;
  service: string;
  title: string;
  subTitle: string;
  description: string;
  whileContacting: string;
  seeServices: string;
  seeProducts: string;
}